move email functions to their own place
